add commit

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -13,7 +13,8 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i;
+	unsigned int i;
+
 	if (array != NULL && size > 0 && action != NULL)
 	{
 	for (i = 0; i < size; i++)
